revert back to use single db for alerts and meta
