+++ src/Controller/SymUrlController.php
@@ -90,7 +90,7 @@ class SymUrlController extends ApiController
      *                 ),
      *                 @SWG\Property(
      *                     property="data",
-     *                     ref=@Model(type=\App\Entity\Ioda\SymUrl::class, groups={"public"})
+     *                     ref=@Model(type=\App\Entity\SymUrl::class, groups={"public"})
      *                 )
      *             )
      *         }
@@ -175,7 +175,7 @@ class SymUrlController extends ApiController
      *                 ),
      *                 @SWG\Property(
      *                     property="data",
-     *                     ref=@Model(type=\App\Entity\Ioda\SymUrl::class, groups={"public"})
+     *                     ref=@Model(type=\App\Entity\SymUrl::class, groups={"public"})
      *                 )
      *             )
      *         }
