+++ src/Entity/OutagesAlert.php
