+++ src/Entity/OutagesSummary.php
